+++ core/src/main/java/com/es/core/dao/StockDao.java
@@ -1,7 +1,5 @@
 package com.es.core.dao;
 
-import com.es.core.model.phone.Stock;
-
 public interface StockDao {
-    int getStock(long phoneId);
+    long getStock(long phoneId);
 }
